+++ src/Failure.php
@@ -3,6 +3,8 @@ declare(strict_types = 1);
 
 namespace Innmind\Validation;
 
+use Innmind\Immutable\Set;
+
 /**
  * @psalm-immutable
  */
@@ -10,10 +12,12 @@ final class Failure
 {
     /**
      * @param non-empty-string $message
+     * @param Set<\UnitEnum> $tags
      */
     private function __construct(
         private KeyPath $path,
         private string $message,
+        private Set $tags,
     ) {
     }
 
@@ -25,7 +29,7 @@ final class Failure
     #[\NoDiscard]
     public static function of(string $message): self
     {
-        return new self(KeyPath::root(), $message);
+        return new self(KeyPath::root(), $message, Set::of());
     }
 
     /**
@@ -36,7 +40,21 @@ final class Failure
     #[\NoDiscard]
     public function under(string $path): self
     {
-        return new self($this->path->under($path), $this->message);
+        return new self(
+            $this->path->under($path),
+            $this->message,
+            $this->tags,
+        );
+    }
+
+    #[\NoDiscard]
+    public function tag(\UnitEnum $tag): self
+    {
+        return new self(
+            $this->path,
+            $this->message,
+            ($this->tags)($tag),
+        );
     }
 
     #[\NoDiscard]
@@ -52,5 +70,14 @@ final class Failure
     public function message(): string
     {
         return $this->message;
+    }
+
+    /**
+     * @return Set<\UnitEnum>
+     */
+    #[\NoDiscard]
+    public function tags(): Set
+    {
+        return $this->tags;
     }
 }